GSM_KEY 사이즈 조정 관련, 버그 수정(복사 오류) , getLogicEnv  Version 2 개발,

+++ src/main/java/com/kt/smartfarm/service/ControlLogicDeviceVO.java
@@ -20,7 +20,7 @@ import lombok.Setter;
 @Setter
 public class ControlLogicDeviceVO {
 	Long id;
-	Long logicId;
+	Integer logicId;
 	Integer deviceNum;
 	String deviceParamName;
 	String deviceType;
